requirements.txt
aiocontextvars==0.2.2
asgiref==3.3.1
contextvars==2.4
Django==3.1.7
googleapis-common-protos==1.52.0
grpcio==1.36.1
immutables==0.15
opentelemetry-api==1.0.0
opentelemetry-exporter-jaeger==1.0.0
opentelemetry-exporter-jaeger-proto-grpc==1.0.0
opentelemetry-exporter-jaeger-thrift==1.0.0
opentelemetry-instrumentation==0.19b0
opentelemetry-instrumentation-dbapi==0.19b0
opentelemetry-instrumentation-django==0.19b0
opentelemetry-instrumentation-grpc==0.19b0
opentelemetry-instrumentation-sqlite3==0.19b0
opentelemetry-instrumentation-urllib==0.19b0
opentelemetry-instrumentation-wsgi==0.19b0
opentelemetry-sdk==1.0.0
opentelemetry-util-http==0.19b0
protobuf==3.15.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
thrift==0.13.0
wrapt==1.12.1
